requirements.txt
streamlit
ydata-profiling
streamlit-pandas-profiling
boto3
pandas
polars
python-dotenv
